@@ -1,19 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 33
+    buildToolsVersion '34.0.0-rc2'
 
     defaultConfig {
         applicationId "com.n1kdo.lotwlook"
-        minSdkVersion 21
-        targetSdkVersion 28
+        minSdkVersion 24
+        targetSdkVersion 33
     }
-    dexOptions {
-        maxProcessCount 4
-        javaMaxHeapSize "5g"
-    }
-
     buildTypes {
         release {
             minifyEnabled true
@@ -21,8 +16,14 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt') // , 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    implementation 'com.android.support:support-compat:28.0.0'
+    implementation 'androidx.core:core:1.9.0'
+//    implementation 'com.google.android.material:material:1.2.0'
 }
+
